pyproject.toml
[project]
name = "flodata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "fastapi>=0.127.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pendulum>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "freezegun>=1.4.0",
    "faker>=22.0.0",
    "testcontainers>=3.7.0",
]
dev = [
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "pre-commit>=3.6.0",
]
performance = [
    "locust>=2.20.0",
    "pytest-benchmark>=4.0.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=api_service",
    "--cov=stream_service",
    "--cov=gcs_download",
    "--cov=run_streamer_cloudrun",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=25",  # Phase 2: api_service.py at 91%, will add stream_service tests next
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    ".venv/*",
    "archive/*",
    "scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
